pyproject.toml
[tool.poetry]
name = "staticpy"
version = "0.0.4"
description = "A static site generator supporting Pandoc and Mathjax for STEM notes, blogpost, portfolio website, and more."
authors = ["Keto Zhang <keto.zhang@gmail.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
pypandoc = "^1.4"
frozen-flask = "^0.15"
werkzeug = "^1.0.0"
python-frontmatter = "^0.5.0"
flask-assets = "^2.0"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
rope = "^0.16.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
